Remove listener
The listener code requires a direct connection to Redis. In the future
access to Redis will be removed in favor of using subscriptions
available in the GraphQL API. This will be much more secure as clients
will not be accessing Redis directly, but instead going through the API
to subscribe to events.