Create mothrpy.yml (#3)
* Create mothrpy.yml

This Action takes the items from here (https://gitlab.rs21.io/mothr/client-libraries/mothrpy/blob/master/.gitlab-ci.yml) and puts them into action syntax - there is a lot repeated here in order to allow for concurrent / dependent jobs, but there may be a better way of doing the same thing using "steps" in one large "build" job.

* Remove unnecessary jobs

With the removal of the listener we no longer need two different sets of
tests. We can also run the linting as part of the tests.

* Rework deploy

* Remove lint dependency

* Remove caching step from deploy

* Test on multiple versions of Python

* Remove Python 3.6

* Change minimum supported Python version to 3.7

* Remove tests for Python 3.9

The is a bug in pylint related to this issue (https://github.com/PyCQA/pylint/issues/3882) that generates a false positive for some type annotations. Removing Python 3.9 from the matrix for now.

* Separate workflows into test and deploy

* Add test status badge

* Upload code coverage report

* Add badge for code coverage

* Use codecov action